requirements.txt
yt-dlp==2025.1.1
youtube-transcript-api==0.6.2
openai-whisper==20231117
whisperx==3.1.1
torch>=2.1
torchaudio>=2.1
ffmpeg-python==0.2.0
faster-whisper==1.0.3
pydantic==2.8.2
PyYAML==6.0.2
tqdm==4.66.4
rich==13.8.1
uvicorn==0.30.1
fastapi==0.111.0
websockets==12.0
numpy==1.26.4
opencv-python==4.10.0.84
webrtcvad==2.0.10
